add tests

+++ logs_and_tests/src/main/java/nix/com/author/AuthorService.java
@@ -13,54 +13,63 @@ public class AuthorService {
     private AuthorDao authorDao = new AuthorDao();
 
     public void create(Author author) {
-        if (!(isAuthorExistByAge(author) || isAuthorExistByName(author))) {
+        LOGGER_INFO.info("start create author: " + author.getId());
+        if (!isParamNull(author)) {
             authorDao.create(author);
         }
+        LOGGER_INFO.info("end create author: " + author.getId());
     }
 
     public void update(Author author) {
-        if (!(isAuthorExistByAge(author) || isAuthorExistByName(author))) {
+        LOGGER_INFO.info("start update author: " + author.getId());
+        if (!isAuthorExistById(author.getId())) {
             authorDao.update(author);
         }
+        LOGGER_INFO.info("end update author: " + author.getId());
     }
 
     public Author[] readAll() {
+        LOGGER_INFO.info("read all authors");
         return authorDao.readAll();
     }
 
     public Author findById(String id) {
-        return authorDao.findById(id);
+        LOGGER_INFO.info("start find author id: " + id);
+        if (!isAuthorExistById(id)) {
+            return authorDao.findById(id);
+        }
+        LOGGER_INFO.info("end find author id: " + id);
+        return null;
     }
 
     public void delete(String id) {
-        authorDao.delete(id);
+        LOGGER_WARN.warn("start delete author: " + id);
+        if (!isAuthorExistById(id)) {
+            authorDao.delete(id);
+        }
+        LOGGER_WARN.warn("end delete author: " + id);
     }
 
-    private boolean isAuthorExistByName(Author author) {
-        if (author.getName() != null) {
-            Author[] authorDaoByName = authorDao.findByName(author.getName());
-            if (authorDaoByName == null) {
+    private boolean isAuthorExistById(String id) {
+        Author CurrentAuthor = authorDao.findById(id);
+        if (CurrentAuthor.getId() != null) {
+            Author[] authorDaoById = authorDao.findByName(CurrentAuthor.getId());
+            if (authorDaoById == null) {
                 return false;
             } else {
-               LOGGER_ERROR.error("this author is exist: " + author.getName());
+                LOGGER_ERROR.info("this author id is exist: " + CurrentAuthor.getId());
             }
         } else {
-            LOGGER_ERROR.error("name can't be empty");
+            LOGGER_ERROR.error("author id can't be empty");
         }
         return true;
     }
 
-    private boolean isAuthorExistByAge(Author author) {
-        if (author.getAge() != null) {
-            Author[] authorDaoByAge = authorDao.findByAge(author.getAge());
-            if (authorDaoByAge == null) {
-                return false;
-            } else {
-                LOGGER_ERROR.error("this author is exist: " + author.getName());
-            }
-        } else {
-            LOGGER_ERROR.error("name can't be empty");
+    private boolean isParamNull(Author author) {
+        if (author.getAge() == null || author.getName() == null) {
+            return true;
         }
-        return true;
+        return false;
     }
+
 }
